+++ test/data/service/qvst_service_test.mocks.dart
@@ -299,4 +299,21 @@ class MockBackendApi extends _i1.Mock implements _i3.BackendApi {
           ),
         )),
       ) as _i4.Future<_i2.HttpResponse<dynamic>>);
+
+  @override
+  _i4.Future<_i2.HttpResponse<dynamic>> getQvstCampaignStatsById(String? id) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #getQvstCampaignStatsById,
+          [id],
+        ),
+        returnValue: _i4.Future<_i2.HttpResponse<dynamic>>.value(
+            _FakeHttpResponse_0<dynamic>(
+          this,
+          Invocation.method(
+            #getQvstCampaignStatsById,
+            [id],
+          ),
+        )),
+      ) as _i4.Future<_i2.HttpResponse<dynamic>>);
 }